refactor: global refactor
