+++ ecommerce-application/src/components/Pages/LoginPage/LoginPage.tsx
@@ -1,6 +1,7 @@
 import { Header } from '../../common/Header/Header';
 import { LoginPageMain } from './LoginPageMain/LoginPageMain';
-import styles from './LoginPage.module.css';
+
+import styles from './LoginPage.module.scss';
 
 export function LoginPage(): React.ReactElement {
   return (
