+++ ecommerce-application/src/components/Pages/MainPage/MainPageMain/MainPageMain.module.scss
@@ -1,10 +1,10 @@
 .main {
   padding: 1rem;
 }
+
 .links {
   display: flex;
   flex-direction: column;
-
   align-items: start;
   row-gap: 1rem;
 }
