+++ ecommerce-application/src/components/Pages/MainPage/MainPageMain/MainPageMain.tsx
@@ -1,4 +1,5 @@
 import { NavLink } from 'react-router-dom';
+
 import styles from './MainPageMain.module.scss';
 
 export function MainPageMain(): React.ReactElement {
