+++ ecommerce-application/src/components/Pages/RegistrationPage/RegistrationPage.module.scss
@@ -5,6 +5,7 @@
   height: 100%;
   margin: 0 auto;
 }
+
 .main {
   width: 100%;
   height: 100%;
