+++ ecommerce-application/src/components/Pages/RegistrationPage/RegistrationPageMain/RegistrationPageMain.module.scss
@@ -8,28 +8,33 @@ $main-color-rose: #51983c;
   align-items: start;
   background: $main_block-background-color;
   height: 100%;
+
   .title {
     font-size: calc(27px + (28 - 27) * ((100vw - 320px) / (1920 - 320)));
     margin-bottom: calc(20px + (30 - 20) * ((100vw - 320px) / (1920 - 320)));
     color: $main-color-rose;
   }
+
   .info_container {
     flex: 1 0 47%;
     display: flex;
     flex-direction: column;
     margin-bottom: 20px;
     margin-right: 10px;
+
     @media screen and (max-width: 650px) {
       flex: 1 0 50%;
     }
+
   }
+
   .info_container:first-child {
     border: 1px solid $main-color-rose;
     border-radius: 10px;
     padding: 10px;
     position: relative;
-    
   }
+
   .address_container {
     border: 1px solid $main-color-rose;
     border-radius: 10px;
@@ -37,11 +42,12 @@ $main-color-rose: #51983c;
     display: flex;
     flex-direction: column;
     position: relative;
-    
   }
+
   .address_container:not(:last-child) {
     margin-bottom: 20px;
   }
+
   .address_title {
     position: absolute;
     top: -10px;
@@ -50,10 +56,12 @@ $main-color-rose: #51983c;
     background-color: $container-background-color;
     color: $main-color-rose;
   }
+
   .form {
     display: flex;
     flex-wrap: wrap;
   }
+
   .submit_button {
     padding: 5px;
     font-size: 18px;
@@ -64,24 +72,29 @@ $main-color-rose: #51983c;
     width: 100%;
     background-color: $main-color-rose;
   }
+
   .text {
     display: inline-block;
     color: grey;
     font-size: 16px;
     margin-right: 5px;
   }
+
   .container {
     display: flex;
     justify-content: space-around;
     align-items: center;
   }
+
   .checkbox_label {
     display: flex;
     flex-direction: row;
   }
+
   .checkbox_input {
     margin-right: 10px;
   }
+
 }
   
 .wrapper {
@@ -90,10 +103,13 @@ $main-color-rose: #51983c;
   padding: 20px 20px;
   margin: 20px 0;
   width: 70%;
+
   @media screen and (max-width: 650px) {
     width: 90%;
   }
+
   .button_login {
+    
     a:hover {
       color: #ff4500
     }
