+++ ecommerce-application/src/components/UI/FormBillingAddressInput/FormBillingAddressInput.module.scss
@@ -3,10 +3,12 @@
   margin-right: 5px;
   align-items: center;
   margin-bottom: 10px;
+  
   .label_block {
     display: inline-block;
     width: 80px;
   }
+  
 }
   
 .input {
