+++ ecommerce-application/src/components/UI/FormCounrtySelect/FormCountrySelect.module.scss
@@ -4,10 +4,12 @@
     align-items: center;
     margin-bottom: 10px;
     font-family: inherit;
+
     .label_block {
         display: inline-block;
         width: 80px;
     }
+    
 }
   
 .select {
