+++ ecommerce-application/src/components/UI/FormPasswordInput/FormPasswordInput.module.scss
@@ -24,10 +24,12 @@
     background-position: center no-repeat;
     margin-bottom: 0px;
     transition: 0.3s;
+
     &:hover {
       transition: 0.3s;
       opacity: 1;
     }
+
   }
 
   .password_button__visible {
@@ -39,6 +41,7 @@
     background: url('/src/assets/img/i-eye-close-black.png') center no-repeat;
     transition: 0.3s;
   }
+  
 }
 
 .input {
