+++ ecommerce-application/src/components/UI/Toast/Toast.tsx
@@ -1,4 +1,5 @@
 import { ToastContainer } from 'react-toastify';
+
 import 'react-toastify/dist/ReactToastify.css';
 
 export function Toast(): React.ReactElement {
