+++ ecommerce-application/src/components/common/Header/Header.module.scss
@@ -31,7 +31,9 @@ li {
 }
 
 @media (width <= 500px) {
+
   ul {
     display: block;
   }
+  
 }
